requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-static==0.0.6
Django==1.11.4
django-appconf==1.0.2
django-geoposition==0.3.1
Django-Select2==5.11.1
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.2
PyJWT==1.5.2
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
static3==0.7.0
urllib3==1.22
whitenoise==3.3.0
